refactor(YALB-110): reorganize existing pages
