+++ components/05-page-examples/miscellaneous/qualtrics-embed.twig
@@ -2,6 +2,7 @@
   {% block page__header %}
     {% include "@molecules/alert/alert.twig" %}
     {% include "@organisms/site-header/site-header.twig" %}
+    {% include "@organisms/menu/breadcrumbs/breadcrumbs.twig" %}
   {% endblock %}
   {% block page__content %}
     {% include "@molecules/page-title/page-title.twig" %}
